UsesDefaultSelectSyntax trait created
Like the UsesAnsiSavepointSyntax, this new trait restrict the standard
implementation of the 'select query'(SelectStatement QueryFragmet). All the
current backens implement this trait, so nothing will change for the end
user.

In custom backends, sometimes we need build the select query from
scracht. Now, we can do this.

And, to allow us ito implement the SelectStatement trait in custom
backends, all SelectStatement field are now public.